Refactor to es6 import/export. Setup development environment - webpack, babel, eslint, hotreload
