+++ .eslintignore
@@ -1,0 +1,2 @@
+src/coverage
+tools/srcServer
