--- bin/www
@@ -1,12 +1,0 @@
-// This will be our application entry. We'll setup our server here.
-const http = require('http');
-const app = require('../server/app'); // The express app we just created
-
-const port = process.env.PORT || 8000;
-app.set('port', port);
-
-const server = http.createServer(app);
-server.listen(port);
-console.log('DocManager running on port: ' + port);
-
-module.exports = app;
