+++ package.json
@@ -4,13 +4,18 @@
   "description": "A beautiful document management system specially made for You!",
   "main": "server.js",
   "scripts": {
+    "start": "npm-run-all --parallel start:dev lint:watch",
     "migrate:test": "NODE_ENV=test sequelize db:migrate",
     "undo:migrate:test": "NODE_ENV=test sequelize db:migrate:undo:all",
     "migrate": "NODE_ENV=development sequelize db:migrate",
     "undo:migrate": "NODE_ENV=development sequelize db:migrate:undo:all",
-    "test": "gulp test",
-    "coverage": "istanbul cover node_modules/mocha/bin/_mocha ./tests/**/*.js",
-    "start:dev": "nodemon bin/www"
+    "coverage": "istanbul cover node_modules/mocha/bin/_mocha \"./src/tests/**/*.js\"",
+    "start:dev": "babel-node -- \"./src/bin/www\"",
+    "open:src": "babel-node tools/srcServer.js",
+    "lint": "node_modules/.bin/esw webpack.config.* src tools",
+    "lint:watch": "npm run lint -- --watch",
+    "test": "mocha --reporter progress tools/testSetup.js \"src/tests/**/*.js\"",
+    "test:watch": "npm run test -- --watch"
   },
   "repository": {
     "type": "git",
@@ -23,23 +28,72 @@
   "author": "Orjiewuru Kingdom Isaac",
   "license": "MIT",
   "dependencies": {
+    "babel-polyfill": "6.8.0",
     "bcrypt-nodejs": "0.0.3",
     "body-parser": "^1.17.1",
+    "babel-register": "6.8.0",
+    "cors": "^2.8.3",
+    "dotenv": "^4.0.0",
     "express": "^4.15.2",
+    "faker": "^4.1.0",
+    "jsdom": "8.5.0",
     "jsonwebtoken": "^7.3.0",
     "morgan": "^1.8.1",
     "path": "^0.12.7",
     "pg": "^6.1.5",
     "pg-hstore": "^2.3.2",
-    "sequelize": "^3.30.4"
+    "sequelize": "^3.30.4",
+    "bootstrap": "3.3.6",
+    "jquery": "2.2.3",
+    "open": "0.0.5",
+    "react": "^15.5.4",
+    "react-dom": "^15.5.4",
+    "react-redux": "4.4.5",
+    "react-router": "2.4.0",
+    "react-router-redux": "4.0.4",
+    "redux": "3.5.2",
+    "redux-thunk": "2.0.1",
+    "toastr": "2.1.2",
+    "webpack": "1.13.0",
+    "webpack-dev-middleware": "1.6.1",
+    "webpack-hot-middleware": "2.10.0"
   },
   "devDependencies": {
-    "cors": "^2.8.3",
-    "dotenv": "^4.0.0",
-    "faker": "^4.1.0",
-    "gulp": "^3.9.1",
-    "gulp-env": "^0.4.0",
-    "gulp-mocha": "^4.3.0",
+    "babel-cli": "6.8.0",
+    "babel-core": "6.8.0",
+    "babel-loader": "6.2.4",
+    "babel-plugin-react-display-name": "2.0.0",
+    "babel-preset-es2015": "6.6.0",
+    "babel-preset-react": "6.5.0",
+    "babel-preset-react-hmre": "1.1.1",
+    "cheerio": "0.22.0",
+    "colors": "1.1.2",
+    "compression": "1.6.1",
+    "cross-env": "1.0.7",
+    "css-loader": "0.23.1",
+    "enzyme": "2.2.0",
+    "eslint": "^3.17.1",
+    "eslint-config-airbnb": "^14.1.0",
+    "eslint-config-airbnb-base": "^11.1.1",
+    "eslint-plugin-import": "^2.2.0",
+    "eslint-plugin-jsdoc": "^2.4.0",
+    "eslint-plugin-jsx-a11y": "^4.0.0",
+    "eslint-plugin-react": "^6.9.0",
+    "eslint-watch": "2.1.11",
+    "eventsource-polyfill": "0.9.6",
+    "expect": "1.19.0",
+    "express": "4.13.4",
+    "extract-text-webpack-plugin": "1.0.1",
+    "file-loader": "0.8.5",
+    "mocha": "2.4.5",
+    "nock": "8.0.0",
+    "npm-run-all": "1.8.0",
+    "react-addons-test-utils": "15.0.2",
+    "redux-immutable-state-invariant": "1.2.3",
+    "redux-mock-store": "1.0.2",
+    "rimraf": "2.5.2",
+    "style-loader": "0.13.1",
+    "url-loader": "0.5.7",
     "istanbul": "^0.4.5",
     "should": "^11.2.1",
     "supertest": "^3.0.0"
