--- server/models/role.js
@@ -1,24 +1,0 @@
-'use strict';
-module.exports = (sequelize, DataTypes) => {
-  var Role = sequelize.define('Role', {
-    title: {
-      allowNull: false,
-      type: DataTypes.STRING,
-      unique: true,
-      validate: {
-        notEmpty: {
-          msg: 'Title cannot be empty'
-        }
-      }
-    }
-  }, {
-    classMethods: {
-      associate: (models) => {
-        Role.hasMany(models.User, {
-          foreignKey: 'RoleId',
-        })
-      }
-    }
-  });
-  return Role;
-};
