--- server/routes/role.routes.js
@@ -1,17 +1,0 @@
-const rolesController = require('../controllers/roles');
-const auth = require('../middleware/authenticate');
-
-const roleRoutes = (router) => {
-  router
-    .route('/roles')
-    .get(rolesController.findAll)
-    .post(rolesController.create);
-
-  router
-    .route('/roles/:roleId')
-    .get(rolesController.findOne)
-    .put(rolesController.update)
-    .delete(rolesController.delete);
-};
-
-module.exports = roleRoutes;
