+++ src/index.html
@@ -1,0 +1,11 @@
+<!DOCTYPE html>
+<html lang="en">
+  <head>
+    <title>Pluralsight Admin</title>
+  </head>
+  <body>
+    <h1>React and Redux in ES6 on Pluralsight</h1>
+    <div id="app"></div>
+    <script src="/bundle.js"></script>
+  </body>
+</html>
