+++ src/index.js
@@ -1,0 +1,2 @@
+/* eslint-disable no-console */
+console.log('Welcome');
