+++ src/server/app.js
@@ -1,0 +1,21 @@
+import express from 'express';
+import logger from 'morgan';
+import bodyParser from 'body-parser';
+import cors from 'cors';
+import routes from './routes/';
+
+// Set up the express app
+const app = express();
+const router = express.Router();
+
+// Parse incoming requests data (https://github.com/expressjs/body-parser)
+app.use(bodyParser.json());
+app.use(bodyParser.urlencoded({ extended: false }));
+// Log requests to the console.
+app.use(logger('dev'));
+app.use(cors());
+
+routes(router);
+app.use('/api/v1', router);
+
+export default app;
