+++ src/server/models/role.js
@@ -1,0 +1,24 @@
+'use strict';
+export default (sequelize, DataTypes) => {
+  const Role = sequelize.define('Role', {
+    title: {
+      allowNull: false,
+      type: DataTypes.STRING,
+      unique: true,
+      validate: {
+        notEmpty: {
+          msg: 'Title cannot be empty'
+        }
+      }
+    }
+  }, {
+    classMethods: {
+      associate: (models) => {
+        Role.hasMany(models.User, {
+          foreignKey: 'RoleId',
+        });
+      }
+    }
+  });
+  return Role;
+};
